moved password init to TeacherController
+ dummy home page
+ redirect to teacher home after login
