+++ site/src/AppBundle/Controller/LoginController.php
@@ -28,7 +28,7 @@ class LoginController extends Controller
             } else {
                 $this->session()->set("user", $t);
                 if ($t->passwordChanged) {
-                    return $this->redirectToRoute("homepage");
+                    return $this->redirectToRoute("teacher_home");
                 } else {
                     return $this->redirectToRoute("init_password");
                 }
@@ -40,35 +40,5 @@ class LoginController extends Controller
                 "teachers"=>$teachers
             ));
         }
-    }
-
-    /**
-     * Reinitialize the teacher password upon first login
-     * or redirect to home page if not the first login
-     *
-     * @Route("/init-password", name="init_password")
-     */
-    public function initPasswordAction()
-    {
-        $user=$this->user();
-        if (!$user) {
-            return $this->redirectToRoute("login");
-        } else if ($user->passwordChanged) {
-            return $this->redirectToRoute("homepage");
-        }
-
-        $request = $this->request();
-        if ($request->getMethod() == "POST") {
-            $password = $request->request->get("password");
-            if (!$password) {
-                throw new \Exception("no password provided ".var_export($password, TRUE));
-            }
-            $user->initPassword($this->db(), $password);
-            return $this->redirectToRoute("homepage");
-        } else {
-            $this->info("%d %d %d", (bool)"0", "" == NULL, "0" == NULL);
-            return $this->render("login/init_password.html.twig");
-        }
-
     }
 }
